# hello-go-project/myfatoorah/cmd/api/docs/swagger.yaml
basePath: /
definitions:
  hello-go-project_myfatoorah_internal_dto.CreatePaymentRequest:
    properties:
      amount:
        example: 100
        type: number
      country_code:
        example: "20"
        type: string
      customer_name:
        description: Optional fields with defaults
        example: Diaa Dawood
        type: string
      email:
        example: diaadawood.mas@gmail.com
        type: string
      expiry_date:
        type: string
      mobile_number:
        example: "1015433746"
        type: string
      order_id:
        example: ORD-2026-001
        type: string
    required:
    - amount
    type: object
  hello-go-project_myfatoorah_internal_dto.CreatePaymentResponse:
    properties:
      invoice_id:
        example: 6646004
        type: integer
      message:
        example: Payment link generated successfully
        type: string
      payment_url:
        example: https://demo.myfatoorah.com/...
        type: string
    type: object
  hello-go-project_myfatoorah_internal_dto.CreateSessionRequest:
    properties:
      amount:
        type: number
      customer_id:
        description: Nullable
        type: string
      order_id:
        description: Nullable
        type: string
      redirection_url:
        type: string
      session_expiry:
        description: Nullable UTC time
        type: string
    required:
    - amount
    type: object
  hello-go-project_myfatoorah_internal_models.Invoice:
    properties:
      created_at:
        type: string
      currency:
        type: string
      customer_email:
        type: string
      customer_name:
        type: string
      id:
        type: string
      order_id:
        description: Internal ID
        type: string
      status:
        description: PENDING, PAID, EXPIRED
        type: string
      total_value:
        type: number
      transactions:
        description: Relationships
        items:
          $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Transaction'
        type: array
      updated_at:
        type: string
    type: object
  hello-go-project_myfatoorah_internal_models.Session:
    properties:
      amount:
        description: Order Details
        type: number
      created_at:
        description: Metadata
        type: string
      currency:
        type: string
      customer_email:
        type: string
      customer_name:
        description: Customer Details (Mapped from the Customer object in response)
        type: string
      customer_reference:
        description: This is your internal "456123789"
        type: string
      encryption_key:
        type: string
      id:
        description: Internal Primary Key
        type: string
      invoice_id:
        type: string
      operation_type:
        description: e.g., "PAY"
        type: string
      session_expiry:
        type: string
      session_id:
        description: MyFatoorah Data
        type: string
      transactions:
        description: Relations
        items:
          $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Transaction'
        type: array
      updated_at:
        type: string
    type: object
  hello-go-project_myfatoorah_internal_models.Transaction:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      error_code:
        type: string
      error_message:
        type: string
      id:
        type: string
      invoice_id:
        description: 'Now a regular index: allows the link creation + multiple payment
          attempts'
        type: string
      invoice_value:
        description: Using decimal for money is safer than float64 in SQL
        type: number
      merchant_id:
        type: string
      mf_invoice_id:
        type: integer
      order_id:
        type: string
      reference:
        description: 'Unique anchor: prevents the same PaymentId from duplicating'
        type: string
      session_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for MyFatoorah.
  title: MyFatoorah API
  version: "1.0"
paths:
  /index.php:
    post:
      consumes:
      - application/json
      description: Receives status updates from MyFatoorah via the route query parameter.
      parameters:
      - description: 'Must be: extension/myfatoorah/payment/myfatoorah.webhook'
        in: query
        name: route
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: 'message: Route not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: MyFatoorah Webhook
      tags:
      - webhooks
  /invoices:
    get:
      description: Retrieves all invoices with their nested session and transaction
        history.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Invoice'
            type: array
        "500":
          description: 'error: database error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all invoices
      tags:
      - invoices
    post:
      consumes:
      - application/json
      description: Creates an internal invoice and generates a MyFatoorah payment
        link.
      parameters:
      - description: Invoice Initiation Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hello-go-project_myfatoorah_internal_dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hello-go-project_myfatoorah_internal_dto.CreatePaymentResponse'
      summary: Initiate Hosted Payment (Invoice/Link)
      tags:
      - invoices
  /invoices/{id}:
    get:
      description: Retrieves an invoice by its UUID, including all associated payment
        sessions and their transaction attempts.
      parameters:
      - description: Invoice UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Invoice'
        "404":
          description: 'error: Invoice not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific invoice
      tags:
      - invoices
  /sessions:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Session'
            type: array
      summary: Get All Sessions with Transactions
      tags:
      - sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Session Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hello-go-project_myfatoorah_internal_dto.CreateSessionRequest'
      produces:
      - application/json
      responses: {}
      summary: Create a new payment session
      tags:
      - sessions
  /transactions:
    get:
      description: Retrieves a list of all payment transactions with search, filter,
        and sorting.
      parameters:
      - description: Search by Order ID or Reference
        in: query
        name: search
        type: string
      - description: Filter by Status (e.g., SUCCESS, FAILED)
        in: query
        name: status
        type: string
      - description: 'Field to sort by (default: created_at)'
        in: query
        name: sort_by
        type: string
      - description: Order direction (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hello-go-project_myfatoorah_internal_models.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Transactions
      tags:
      - transactions
swagger: "2.0"